refactor: move entity erros to domain folder
