+++ src/domain/errors/empty-properties/InvalidEmptyID.ts
