+++ tests/CPF.spec.ts
@@ -1,5 +1,5 @@
 import { CPF } from '@/CPF'
-import { InvalidCPFerror } from '@/errors'
+import { InvalidCPFerror } from '@/domain/errors'
 
 const makeSut = (cpf: string): CPF => {
   return new CPF(cpf)
